Revert "Reverting to int form"
This reverts commit 6774c785d8aa52b6f73e54b12071681aa498c0f3.

+++ tempsensor_v1/src/config.h
@@ -434,7 +434,7 @@ typedef struct {
 	uint32_t lastSeek;
 
 	// GSM or GPRS
-	int network_mode;
+	uint8_t network_mode;
 	NETWORK_SERVICE net_service[2];
 
 	SAFEBOOT_STATUS safeboot;
